Use u32 sort key for quad fitting angle sort
Convert Pt.slope from f32 to u32 (bit pattern of non-negative f32) and
use sort_unstable_by_key instead of sort_unstable_by with total_cmp.
Since slope values are always in [0, 4) (non-negative, non-NaN in normal
operation), IEEE 754 bit patterns preserve ordering as plain integers.
This eliminates the total_cmp closure overhead per comparison.

+++ apriltag/src/detect/cluster.rs
@@ -12,8 +12,8 @@ pub struct Pt {
     pub gx: i16,
     /// Gradient direction y component.
     pub gy: i16,
-    /// Angular sort key for quad fitting.
-    pub slope: f32,
+    /// Angular sort key for quad fitting (bit pattern of non-negative f32).
+    pub slope: u32,
 }
 
 /// A cluster of edge points sharing the same component-pair boundary.
@@ -73,7 +73,7 @@ pub fn gradient_clusters(
                         y: (2 * $y as i32 + $dy) as u16,
                         gx,
                         gy,
-                        slope: 0.0,
+                        slope: 0,
                     };
                     $pairs.push((key, pt));
                     true
